Changed linter to ESLint
JSHint is not as strict with rules as I would like. Changing to
ESLint has given a stricter and more configurable linting solution.
Because of the linter change, there was some changes in most of the
files as the spacing is more strictly enforced now. It did catch a
few bugs and useless variables. Seeing imporvement already.