Connected to Spotify API
The search method is set up to hit the Spotify API and return a
json object containing the results. This is a great first step in
connecting to the API. I will be pushing this live on staging so
that we can connect the front end to this for further testing.