Merged In React Frontend
Was able to merge the React frontend into the api and everything
still works and loads as expected. There have been some sweeping
changes in the routing paths for the api calls. I also did my best
to make sure that everything worked. Tests for the API are good,
but the frontend is still not tested as far as I would like. There
will likely be more important changes made as we continue to work
on the merged frontend.