Fixes #18 Token Refresh Issue
Upon closer inspection, the issue stemmed from the order of args in
the functions that were being wrapped. The search function had user
listed second, and the get playlists function only had one param, the
user. This caused issues because the placement of user in the args was
not consistent. THe solution is to make sure that user is always
first in the args list.