Updated remote_test mocha reporter to min
This cutrs down on the noise in the output logs when running tests
remotely. It should still show information on failed tests when they
occur, but will nort show every passed test. Also remove the build
folder from git and added it to the .gitignore file as it is a build
generated file and should not have been stored anyway.