Updated Spotify GET playlist route
The route has been updated to get a playlist by the url. This was
needed to fix an issue where in the service would try and fetch a
playlist created by another user as if it were created by the logged
in user. This fix should handle that better. The SPotify service
was also updated to handle 502 bad gateway issues that popped up.